pyproject.toml
[project]
name = "zirak"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "google-cloud-aiplatform>=1.87.0",
    "google-genai>=1.9.0",
    "mastodon-py>=2.0.1",
    "pydantic-settings>=2.8.1",
    "pydantic-yaml>=1.4.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
